requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
tqdm
mlflow
dvc
dvc[s3]
PyYAML
Pillow